Update Tinkers Construct to 1.12, Currently there are some bugs; Tools/Patterns/Cast's show up as missing textures, Integration is currently broken, and the JEI plugin doesn't function correctly when showing the table recipes.
Todo:
Crafting Stations need support for the Crafting Book
Ensure all crafting recipes work as expected

+++ src/main/java/slimeknights/tconstruct/gadgets/client/RenderFancyItemFrame.java
@@ -7,9 +7,9 @@ import net.minecraft.client.renderer.RenderItem;
 import net.minecraft.client.renderer.block.model.IBakedModel;
 import net.minecraft.client.renderer.block.model.ModelManager;
 import net.minecraft.client.renderer.entity.Render;
+import net.minecraft.client.renderer.entity.RenderItemFrame;
 import net.minecraft.client.renderer.entity.RenderManager;
 import net.minecraft.client.renderer.texture.TextureMap;
-import net.minecraft.client.renderer.tileentity.RenderItemFrame;
 import net.minecraft.entity.item.EntityItemFrame;
 import net.minecraft.init.Items;
 import net.minecraft.util.math.BlockPos;
@@ -43,9 +43,9 @@ public class RenderFancyItemFrame extends RenderItemFrame {
 
     GlStateManager.pushMatrix();
     BlockPos blockpos = entity.getHangingPosition();
-    double d0 = (double) blockpos.getX() - entity.posX + x;
-    double d1 = (double) blockpos.getY() - entity.posY + y;
-    double d2 = (double) blockpos.getZ() - entity.posZ + z;
+    double d0 = blockpos.getX() - entity.posX + x;
+    double d1 = blockpos.getY() - entity.posY + y;
+    double d2 = blockpos.getZ() - entity.posZ + z;
     GlStateManager.translate(d0 + 0.5D, d1 + 0.5D, d2 + 0.5D);
     GlStateManager.rotate(180.0F - entity.rotationYaw, 0.0F, 1.0F, 0.0F);
     this.renderManager.renderEngine.bindTexture(TextureMap.LOCATION_BLOCKS_TEXTURE);
@@ -73,7 +73,7 @@ public class RenderFancyItemFrame extends RenderItemFrame {
     this.renderItem(entity);
     GlStateManager.popMatrix();
     assert entity.facingDirection != null;
-    this.renderName(entity, x + (double) ((float) entity.facingDirection.getFrontOffsetX() * 0.3F), y - 0.25D, z + (double) ((float) entity.facingDirection.getFrontOffsetZ() * 0.3F));
+    this.renderName(entity, x + entity.facingDirection.getFrontOffsetX() * 0.3F, y - 0.25D, z + entity.facingDirection.getFrontOffsetZ() * 0.3F);
   }
 
   private static class Factory implements IRenderFactory<EntityFancyItemFrame> {
